add js <script/> tags to bottom of base.html
on the last push jquery datatable finally worked on exercise.html! the
issue was script wasn't in between the blocks on exercise pg (as it
inherits from base.html) + the tags were all in the <head/> of
base.html. this push keeps all .css links in base.html <head/> but moves
.js scripts to end of base.html <body/>