reorganize session add and commit
before this function depended on the user having started a workout,
therefore a workout object was started in the database to add values to.
this time it's separately adding half created workout objects to the
session but only committing when finish workout is pressed.
also removed dash from requirements.txt lol