add validation to finish workout
tried copy pasting this script into browser but was getting 500 server
error, so the conditionals weren't working... however, also changed
increment from 1 to 2 on exerciseNumber and page didn't seem to respond
to change. so maybe changes can only be made to existing scripts?
confused exactly what the limits of testing this way are